Refine sample smart contract and client application

+++ contracts/asset-tx-typescript/src/assetTransfer.ts
@@ -3,20 +3,41 @@
  */
 
 import { Context, Contract, Info, Returns, Transaction } from 'fabric-contract-api';
+import { KeyEndorsementPolicy } from 'fabric-shim';
 import stringify from 'json-stringify-deterministic'; // Deterministic JSON.stringify()
 import sortKeysRecursive from 'sort-keys-recursive';
 import { TextDecoder } from 'util';
-import { Asset } from './asset';
+import { Asset, newAsset } from './asset';
 
 const utf8Decoder = new TextDecoder();
+
+function unmarshal(bytes: Uint8Array | string): object {
+    const json = typeof bytes === 'string' ? bytes : utf8Decoder.decode(bytes);
+    const parsed: unknown = JSON.parse(json);
+    if (parsed === null || typeof parsed !== 'object') {
+        throw new Error(`Invalid JSON type (${typeof parsed}): ${json}`);
+    }
+
+    return parsed;
+}
 
 function marshal(o: object): Buffer {
     return Buffer.from(toJSON(o));
 }
 
 function toJSON(o: object): string {
-    // we insert data in alphabetic order using 'json-stringify-deterministic' and 'sort-keys-recursive'
+    // Insert data in alphabetic order using 'json-stringify-deterministic' and 'sort-keys-recursive'
     return stringify(sortKeysRecursive(o));
+}
+
+function newMemberPolicy(...orgs: string[]): KeyEndorsementPolicy {
+    const policy = new KeyEndorsementPolicy();
+    policy.addOrgs('MEMBER', ...orgs);
+    return policy;
+}
+
+function hasWritePermission(ctx: Context, owner: string): boolean {
+    return owner === ctx.clientIdentity.getID();
 }
 
 @Info({title: 'AssetTransfer', description: 'Smart contract for trading assets'})
@@ -25,16 +46,23 @@ export class AssetTransferContract extends Contract {
      * CreateAsset issues a new asset to the world state with given details.
      */
     @Transaction()
-    public async CreateAsset(ctx: Context, assetData: string): Promise<void> {
-        const asset = Asset.unmarshal(assetData);
+    async CreateAsset(ctx: Context, assetJson: string): Promise<void> {
+        const state = Object.assign(unmarshal(assetJson), {
+            Owner: ctx.clientIdentity.getID(),
+        });
+        const asset = newAsset(state);
+
         const exists = await this.AssetExists(ctx, asset.ID);
         if (exists) {
             throw new Error(`The asset ${asset.ID} already exists`);
         }
 
-        // we insert data in alphabetic order using 'json-stringify-deterministic' and 'sort-keys-recursive'
         const assetBytes = marshal(asset);
         await ctx.stub.putState(asset.ID, assetBytes);
+
+        const policy = newMemberPolicy(ctx.clientIdentity.getMSPID());
+        await ctx.stub.setStateValidationParameter(asset.ID, policy.getPolicy());
+
         ctx.stub.setEvent('CreateAsset', assetBytes);
     }
 
@@ -42,7 +70,7 @@ export class AssetTransferContract extends Contract {
      * ReadAsset returns an existing asset stored in the world state.
      */
     @Transaction(false)
-    public async ReadAsset(ctx: Context, id: string): Promise<string> {
+    async ReadAsset(ctx: Context, id: string): Promise<string> {
         const assetBytes = await this.#readAsset(ctx, id);
         return utf8Decoder.decode(assetBytes);
     }
@@ -61,22 +89,31 @@ export class AssetTransferContract extends Contract {
      * the asset ID.
      */
     @Transaction()
-    public async UpdateAsset(ctx: Context, assetData: string): Promise<void> {
-        const assetUpdate = JSON.parse(assetData) as Partial<Asset>;
+    async UpdateAsset(ctx: Context, assetJson: string): Promise<void> {
+        const assetUpdate: Partial<Asset> = unmarshal(assetJson);
         if (assetUpdate.ID === undefined) {
             throw new Error('No asset ID specified');
         }
 
         const existingAssetBytes = await this.#readAsset(ctx, assetUpdate.ID);
-        const existingAsset = Asset.unmarshal(existingAssetBytes);
+        const existingAsset = newAsset(unmarshal(existingAssetBytes));
 
-        const updatedAsset = Object.assign(existingAsset, assetUpdate, {
+        if (!hasWritePermission(ctx, existingAsset.Owner)) {
+            throw new Error('Only owner can update assets');
+        }
+
+        const updatedState = Object.assign(existingAsset, assetUpdate, {
             Owner: existingAsset.Owner, // Must transfer to change owner
         });
+        const updatedAsset = newAsset(updatedState);
 
         // overwriting original asset with new asset
         const updatedAssetBytes = marshal(updatedAsset);
         await ctx.stub.putState(updatedAsset.ID, updatedAssetBytes);
+
+        const policy = newMemberPolicy(ctx.clientIdentity.getMSPID());
+        await ctx.stub.setStateValidationParameter(updatedAsset.ID, policy.getPolicy());
+
         ctx.stub.setEvent('UpdateAsset', updatedAssetBytes);
     }
 
@@ -84,10 +121,16 @@ export class AssetTransferContract extends Contract {
      * DeleteAsset deletes an asset from the world state.
      */
     @Transaction()
-    public async DeleteAsset(ctx: Context, id: string): Promise<void> {
+    async DeleteAsset(ctx: Context, id: string): Promise<void> {
         const assetBytes = await this.#readAsset(ctx, id); // Throws if asset does not exist
+        const asset = newAsset(unmarshal(assetBytes));
+
+        if (!hasWritePermission(ctx, asset.Owner)) {
+            throw new Error('Only owner can delete assets');
+        }
 
         await ctx.stub.deleteState(id);
+
         ctx.stub.setEvent('DeletaAsset', assetBytes);
     }
 
@@ -96,28 +139,32 @@ export class AssetTransferContract extends Contract {
      */
     @Transaction(false)
     @Returns('boolean')
-    public async AssetExists(ctx: Context, id: string): Promise<boolean> {
-        const assetJSON = await ctx.stub.getState(id);
-        return assetJSON && assetJSON.length > 0;
+    async AssetExists(ctx: Context, id: string): Promise<boolean> {
+        const assetJson = await ctx.stub.getState(id);
+        return assetJson?.length > 0;
     }
 
     /**
-     * TransferAsset updates the owner field of asset with the specified ID in the world state, and returns the old
-     * owner.
+     * TransferAsset updates the owner field of asset with the specified ID in the world state.
      */
     @Transaction()
-    public async TransferAsset(ctx: Context, id: string, newOwner: string): Promise<string> {
-        const assetString = await this.ReadAsset(ctx, id);
-        const asset = JSON.parse(assetString) as Asset;
+    async TransferAsset(ctx: Context, id: string, newOwner: string, newOwnerOrg: string): Promise<void> {
+        const assetString = await this.#readAsset(ctx, id);
+        const asset = newAsset(unmarshal(assetString));
 
-        const oldOwner = asset.Owner;
+        if (!hasWritePermission(ctx, asset.Owner)) {
+            throw new Error('Only owner can transfer assets');
+        }
+
         asset.Owner = newOwner;
 
         const assetBytes = marshal(asset);
         await ctx.stub.putState(id, assetBytes);
-        ctx.stub.setEvent('TransferAsset', assetBytes);
 
-        return oldOwner;
+        const policy = newMemberPolicy(newOwnerOrg);
+        await ctx.stub.setStateValidationParameter(id, policy.getPolicy());
+
+        ctx.stub.setEvent('TransferAsset', assetBytes);
     }
 
     /**
@@ -125,7 +172,7 @@ export class AssetTransferContract extends Contract {
      */
     @Transaction(false)
     @Returns('string')
-    public async GetAllAssets(ctx: Context): Promise<string> {
+    async GetAllAssets(ctx: Context): Promise<string> {
         // range query with empty string for startKey and endKey does an open-ended query of all assets in the chaincode namespace.
         const iterator = await ctx.stub.getStateByRange('', '');
 
@@ -133,7 +180,7 @@ export class AssetTransferContract extends Contract {
         for (let result = await iterator.next(); !result.done; result = await iterator.next()) {
             const assetBytes = result.value.value;
             try {
-                const asset = Asset.unmarshal(assetBytes);
+                const asset = newAsset(unmarshal(assetBytes));
                 assets.push(asset);
             } catch (err) {
                 console.log(err);
